+++ Source/OCMock/OCObserverMockObject.m
@@ -64,7 +64,7 @@
 
 - (void)handleNotification:(NSNotification *)aNotification
 {
-	int i, limit;
+	NSUInteger i, limit;
 	
 	limit = expectationOrderMatters ? 1 : [recorders count];
 	for(i = 0; i < limit; i++)
